Refactor song reading into helper
